==> scripts/03_TNF_welch.Rmd <==
---
title: "TNF_welch"
author: "Elsie Jacobson"
date: "8/8/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(Vennerable)
library(gplots)
library(scales)
library(ggrepel)
library(extrafont)
library(broom)

ggformat <- theme_classic(base_family = "Arial") + theme(text = element_text(size = 20))
```

```{r}
pro_all <- read.csv("../results/Pro_TNF_all.csv", row.names=1)
pro_sig <- read.csv("../results/Pro_TNF_sig.csv", row.names=1)
gran_all <- read.csv("../results/Gran_TNF_all.csv", row.names=1)
gran_sig <- read.csv("../results/Gran_TNF_sig.csv", row.names=1)

TPA_all <- read.csv("../results/tpa_all.csv", row.names=1)
TPA_sig <- read.csv("../results/tpa_sig.csv", row.names=1)
RA_all <- read.csv("../results/atra_all.csv", row.names=1)
RA_sig <- read.csv("../results/atra_sig.csv", row.names=1)
```

Make long form dataset of DGE

```{r}
all_long <- pro_all %>%
  mutate(sample="pro+TNF") %>%
  bind_rows(gran_all) %>%
  mutate(sample=case_when(is.na(sample)==TRUE~"gran+TNF", TRUE~sample)) %>%
  bind_rows(RA_all) %>%
  mutate(sample=case_when(is.na(sample)==TRUE~"pro+ATRA", TRUE~sample)) %>%
  bind_rows(TPA_all) %>%
  mutate(sample=case_when(is.na(sample)==TRUE~"pro+TPA", TRUE~sample)) %>%
  distinct() %>%
  mutate(sig=case_when(padj<0.05~"sig",TRUE~"NS")) %>%
  mutate(change=case_when(abs(log2FoldChange)>2~"log2FC>2",
                          abs(log2FoldChange)>1~"log2FC>1",
                          TRUE~"log2FC<=1"))
write.csv(all_long, "../results/TNF_diff_longform.csv")
```

Compile dataset of DGE in both comparisons

```{r}
p <- paste(colnames(pro_all), "p",sep="_")
g <- paste(colnames(gran_all), "g",sep="_")
r <- paste(colnames(RA_all), "r",sep="_")
t <- paste(colnames(TPA_all), "t",sep="_")

colnames(pro_all) <- p
colnames(gran_all) <- g
colnames(RA_all) <- r
colnames(TPA_all) <- t

all <- pro_all %>%
  full_join(gran_all, by=c("ens_gene_p"="ens_gene_g")) %>%
  full_join(RA_all, by=c("ens_gene_p"="ens_gene_r")) %>%
  full_join(TPA_all, by=c("ens_gene_p"="ens_gene_t")) %>%
  distinct()

all_sig <- all %>%
  mutate(sig_p=ifelse(padj_p<0.05,"p_sig","p_NS")) %>%
  mutate(sig_g=ifelse(padj_g<0.05,"g_sig","g_NS")) %>%
  mutate(sig_r=ifelse(padj_r<0.05,"r_sig","r_NS")) %>%
  mutate(sig_t=ifelse(padj_t<0.05,"t_sig","t_NS")) %>%
  filter(sig_p=="p_sig" | sig_g=="g_sig" | sig_r=="r_sig" | sig_t=="t_sig") %>%
  na.omit() %>%
  distinct() %>%
  dplyr::select(ens_gene=ens_gene_p, external_gene_name=external_gene_name_p, baseMean_proTNF=baseMean_p, log2FoldChange_proTNF=log2FoldChange_p, padj_proTNF=padj_p, baseMean_granTNF=baseMean_g, log2FoldChange_granTNF=log2FoldChange_g, padj_granTNF=padj_g, baseMean_ATRA=baseMean_r, log2FoldChange_ATRA=log2FoldChange_r, padj_ATRA=padj_r, baseMean_TPA=baseMean_t, log2FoldChange_TPA=log2FoldChange_t, padj_TPA=padj_t)
write.csv(all_sig, "../results/HL60_ATRA_TPA_TNF_sig.csv")

write.csv(all, "../results/HL60_ATRA_TPA_TNF_all.csv")

universe <- all$ens_gene_p
write.csv(universe, "../results/gene_universe_tnf.csv", row.names = FALSE)
```

Supp1***************

```{r}
 supp1 <- all_sig %>%
  mutate(TNF_Category=case_when(padj_proTNF<0.05 & padj_granTNF>0.05~"Pro-specific TNF",
                               padj_proTNF>0.05 & padj_granTNF<0.05~"Gran-specific TNF",
                               padj_proTNF<0.05 & padj_granTNF<0.05 & log2FoldChange_proTNF*log2FoldChange_granTNF>=0~"Conserved TNF",
                               padj_proTNF<0.05 & padj_granTNF<0.05 & log2FoldChange_proTNF*log2FoldChange_granTNF<0~"Opposite TNF")) %>%
  mutate(Diff_category=case_when(padj_ATRA<0.05 & padj_TPA>0.05~"ATRA-specific",
                               padj_ATRA>0.05 & padj_TPA<0.05~"TPA-specific",
                               padj_ATRA<0.05 & padj_TPA<0.05 & log2FoldChange_ATRA*log2FoldChange_TPA>=0~"Conserved diff",
                               padj_ATRA<0.05 & padj_TPA<0.05 & log2FoldChange_ATRA*log2FoldChange_TPA<0~"Opposite diff"))
write.csv(supp1, "../results/Supplementary_table_1.csv")
```

```{r}
summary_FC <- all_long %>%
  group_by(sample) %>%
  filter(padj<0.05) %>%
  summarise(mean=mean(log2FoldChange),median=median(log2FoldChange),sd=sd(log2FoldChange),
            min=min(log2FoldChange),max=max(log2FoldChange),p_25=quantile(log2FoldChange, probs=0.25),
            p_75=quantile(log2FoldChange, probs=0.75))
```

Volcano plots

```{r}
ggplot(all_long, aes(log2FoldChange, -log(padj), colour=padj<0.05)) + geom_point(alpha=0.5) + facet_wrap(~sample) + ggformat + scale_colour_viridis_d(option="B", begin=0.2, end=0.8) + labs(colour="DE")
ggsave("../results/all_volcanoes.pdf", useDingbats=FALSE)
```

```{r}
table(pro_sig$log2FoldChange>0)
prop.test(c(980,1312),c(2292,2292))

pt <- ggplot(pro_all, aes(log2FoldChange_p, -log(padj_p), colour=padj_p<0.05, label=external_gene_name_p)) + geom_point(alpha=0.5, size=1) + scale_colour_manual(values=c("black","#cc2a2a")) + ggformat + labs(x="log2 Fold Change pro +/- TNF", y="-log adjusted p value", colour="Significant")
ggsave("../results/volcano_pro_TNF.pdf", useDingbats=FALSE)
```

```{r}
table(gran_sig$log2FoldChange>0)
prop.test(c(3597,3809),c(7406,7406))

gt <- ggplot(gran_all, aes(log2FoldChange_g, -log(padj_g), colour=padj_g<0.05, label=external_gene_name_g)) + geom_point(alpha=0.5) + scale_colour_manual(values=c("black","#cc2a2a")) + ggformat  + labs(x="log2 Fold Change gran +/- TNF", y="-log adjusted p value", colour="Significant")
ggsave("../results/volcano_gran_TNF.pdf", useDingbats=FALSE)
```

```{r}
plot_grid(pt, gt)
ggsave("../results/volcano_pro_gran_TNF.tiff", width=14)
```

###Venn diagram of the intersects

```{r}
venrbl <- NULL
venrbl$pro <- as.character(all$ens_gene_p[all$padj_p<0.05])
venrbl$gran <- as.character(all$ens_gene_p[all$padj_g<0.05])
venrbl$ra <- as.character(all$ens_gene_p[all$padj_r<0.05])
venrbl$tpa <- as.character(all$ens_gene_p[all$padj_t<0.05])
str(venrbl)

V_TNF <- Venn(venrbl)
V_TNF
plot(V_TNF, type="ellipses")
```

HL60 + TNF vs HL60-RA + TNF

```{r}
ven_pro_gran <- NULL
ven_pro_gran$pro <- unique(as.character(all$ens_gene_p[all$padj_p<0.05]))
ven_pro_gran$gran <- unique(as.character(all$ens_gene_p[all$padj_g<0.05]))

V_pro_gran <- Venn(ven_pro_gran)
pdf("../results/Venn_pro_gran.pdf")
plot(V_pro_gran, show=list(Faces=FALSE,DarkMatter=FALSE))
dev.off()
```

```{r}
pro_gran <- all_sig %>%
  filter(padj_proTNF<0.05 | padj_granTNF<0.05) %>%
  mutate(Significant=case_when(padj_proTNF<0.05 & padj_granTNF<0.05~"Both",
                               padj_proTNF<0.05 & padj_granTNF>0.05~"Pro-specific",
                               padj_proTNF>0.05 & padj_granTNF<0.05~"Gran-specific")) %>%
  mutate(Significant=factor(Significant, levels=c("Both","Pro-specific","Gran-specific")))

ggplot(pro_gran, aes(log2FoldChange_proTNF, log2FoldChange_granTNF)) + geom_point(alpha=0.5) + ggformat + geom_hline(yintercept=0) + geom_vline(xintercept = 0) + geom_smooth(method="lm") + facet_wrap(~Significant, scales="free")
ggsave("../results/pro_gran_log2FC.pdf")

both_tnf <- tidy(lm(log2FoldChange_granTNF~log2FoldChange_proTNF, data=subset(all_sig, padj_proTNF<0.05 & padj_granTNF<0.05)))
both_tnf$model <- "Pro and Gran"
pro <- tidy(lm(log2FoldChange_granTNF~log2FoldChange_proTNF, data=subset(all_sig, padj_proTNF<0.05 & padj_granTNF>0.05)))
pro$model <- "Pro only"
gran <- tidy(lm(log2FoldChange_granTNF~log2FoldChange_proTNF, data=subset(all_sig, padj_proTNF>0.05 & padj_granTNF<0.05)))
gran$model <- "Gran only"

TNF_lm <- rbind(both_tnf,pro,gran)
```


```{r}
pro_genes <- unique(as.character(all$ens_gene_p[all$padj_p<0.05]))
gran_genes <- unique(as.character(all$ens_gene_p[all$padj_g<0.05]))
pro_all <- unique(as.character(all$ens_gene_p))
gran_all <- unique(as.character(all$ens_gene_p))
N <- 1000

inter.r <- length(intersect(pro_genes,gran_genes))

getSimIntersect <- function(pro_genes,gran_genes,pro_all,gran_all){
  gA.sim <- sample(pro_all,length(pro_genes))
  gB.sim <- sample(gran_all,length(gran_genes))
  return(length(intersect(gA.sim,gB.sim)))
}

inter.sim <- replicate(N,getSimIntersect(pro_genes,gran_genes,pro_all,gran_all))

plot(density(inter.sim))
summary(inter.sim)
quantile(inter.sim)
1662/592
```

HL60 + TNF vs RA vs TPA - TNF1

```{r}
ven_pro <- NULL
ven_pro$pro <- as.character(all$ens_gene_p[all$padj_p<0.05])
ven_pro$ra <- as.character(all$ens_gene_p[all$padj_r<0.05])
ven_pro$tpa <- as.character(all$ens_gene_p[all$padj_t<0.05])


V_pro <- Venn(ven_pro)
V_pro
pdf("../results/Venn_pro.pdf")
plot(V_pro, show=list(Faces=FALSE,DarkMatter=FALSE))
dev.off()
```

```{r}
tnf <- unique(as.character(all$ens_gene_p[all$padj_p<0.05]))
ra <- unique(as.character(all$ens_gene_p[all$padj_r<0.05]))
tpa <- unique(as.character(all$ens_gene_p[all$padj_t<0.05]))
all_genes <- unique(as.character(all$ens_gene_p))
N <- 1000

inter.r <- length(intersect(intersect(tnf,ra),tpa))

getSimIntersect <- function(tnf,ra,tpa,all_genes){
  gA.sim <- sample(all_genes,length(tnf))
  gB.sim <- sample(all_genes,length(ra))
  gC.sim <- sample(all_genes,length(tpa))
  return(length(intersect(intersect(gA.sim,gB.sim),gC.sim)))
}

inter.sim <- replicate(N,getSimIntersect(tnf,ra,tpa,all_genes))
p <- sum(inter.sim>=inter.r)/N
plot(density(inter.sim))
summary(inter.sim)
inter.r
quantile(inter.sim)
inter.r/mean(inter.sim)
```

HL60-RA + TNF vs RA vs TPA

```{r}
ven_gran <- NULL
ven_gran$gran <- as.character(gran_sig$ens_gene)
ven_gran$ra <- as.character(RA_sig$ens_gene)
ven_gran$tpa <- as.character(TPA_sig$ens_gene)

V_gran <- Venn(ven_gran)
V_gran
pdf("../results/Venn_gran.pdf")
plot(V_gran, show=list(Faces=FALSE,DarkMatter=FALSE))
dev.off()
```

```{r}
tnf <- unique(as.character(all$ens_gene_p[all$padj_g<0.05]))
ra <- unique(as.character(all$ens_gene_p[all$padj_r<0.05]))
tpa <- unique(as.character(all$ens_gene_p[all$padj_t<0.05]))
all_genes <- unique(as.character(all$ens_gene_p))
N <- 1000

inter.r <- length(intersect(intersect(tnf,ra),tpa))

getSimIntersect <- function(tnf,ra,tpa,all_genes){
  gA.sim <- sample(all_genes,length(tnf))
  gB.sim <- sample(all_genes,length(ra))
  gC.sim <- sample(all_genes,length(tpa))
  return(length(intersect(intersect(gA.sim,gB.sim),gC.sim)))
}

inter.sim <- replicate(N,getSimIntersect(tnf,ra,tpa,all_genes))
p <- sum(inter.sim>=inter.r)/N
plot(density(inter.sim))
summary(inter.sim)
inter.r
quantile(inter.sim)
inter.r/mean(inter.sim)
```

```{r}
#pro 331/2292 not in diff
#gran 1555/7406 not in diff

prop.test(c(331,1555),c(2292,7406))
```

Pro+TNF vs RA

```{r}
pro_ra <- subset(all_sig,sig_r=="r_sig" & sig_p=="p_sig")

summary(lm(log2FoldChange_p~log2FoldChange_r, data=pro_ra))

ggplot(pro_ra, aes(log2FoldChange_r, log2FoldChange_p)) + geom_point(alpha=0.5) + ggformat + geom_hline(yintercept=0) + geom_vline(xintercept = 0) + geom_smooth(method="lm")
ggsave("../results/pro_vs_RA_log2FC.pdf")
```

Gran+TNF vs RA

```{r}
gran_ra <- subset(all_sig,sig_r=="r_sig" & sig_g=="g_sig")

summary(lm(log2FoldChange_g~log2FoldChange_r, data=gran_ra))

ggplot(gran_ra, aes(log2FoldChange_r, log2FoldChange_g)) + geom_point(alpha=0.5) + ggformat + geom_hline(yintercept=0) + geom_vline(xintercept = 0) + geom_smooth(method="lm")
ggsave("../results/gran_vs_RA_log2FC.pdf")
```

Pro+TNF vs TPA

```{r}
pro_tpa <- subset(all_sig,sig_t=="t_sig" & sig_p=="p_sig")

summary(lm(log2FoldChange_p~log2FoldChange_t, data=pro_tpa))

ggplot(pro_tpa, aes(log2FoldChange_t, log2FoldChange_p)) + geom_point(alpha=0.5) + ggformat + geom_hline(yintercept=0) + geom_vline(xintercept = 0) + geom_smooth(method="lm")
ggsave("../results/pro_vs_TPA_log2FC.pdf")
```

Gran+TNF vs TPA

```{r}
gran_tpa <- subset(all_sig,sig_t=="t_sig" & sig_g=="g_sig")

summary(lm(log2FoldChange_g~log2FoldChange_t, data=gran_tpa))

ggplot(gran_tpa, aes(log2FoldChange_t, log2FoldChange_g)) + geom_point(alpha=0.5) + ggformat + geom_hline(yintercept=0) + geom_vline(xintercept = 0) + geom_smooth(method="lm")
ggsave("../results/gran_vs_TPA_log2FC.pdf")
```

TNF level 1 - pro+TNF, ATRA, TPA

```{r}
TNF1 <- all_sig %>%
  mutate(allsig=case_when(sig_t=="t_sig" & sig_p=="p_sig" & sig_r=="r_sig"~"allsig",
                          TRUE~"NS")) %>%
  mutate(allup=case_when(log2FoldChange_t>0 & log2FoldChange_p>0 & log2FoldChange_r>0 ~"up",TRUE~"not_up")) %>%
  mutate(alldown=case_when(log2FoldChange_t<0  & log2FoldChange_p<0 & log2FoldChange_r<0 ~"down",TRUE~"not_down")) %>%
  mutate(same_dir=case_when(allup=="up" | alldown=="down"~"same",TRUE~"diff")) %>%
  filter(allsig=="allsig", same_dir=="same") %>%
  distinct()

write.csv(TNF1, "../results/TNF_response_level1.csv")
```

TNF level 2 - gran+TNF, TPA excluding TNF1

```{r}
TNF2 <- all_sig %>%
  mutate(allsig=case_when(sig_t=="t_sig" & sig_g=="g_sig"~"allsig", TRUE~"NS")) %>%
  mutate(allup=case_when(log2FoldChange_t>0 & log2FoldChange_g>0 ~"up",TRUE~"not_up")) %>%
  mutate(alldown=case_when(log2FoldChange_t<0 & log2FoldChange_g<0 ~"down",TRUE~"not_down")) %>%
  mutate(same_dir=case_when(allup=="up" | alldown=="down"~"same",TRUE~"diff")) %>%
  filter(allsig=="allsig", same_dir=="same", !(ens_gene_p %in% TNF1$ens_gene_p)) %>%
  distinct()

write.csv(TNF2, "../results/TNF_response_level2.csv")
```

prop test - TNF1 genes

```{r}
#pro vs ra
prop.test(c(660,660),c(2292,11255))
#ra vs tpa
prop.test(c(660,660),c(11255,13177))
```

prop test - TNF2 genes

```{r}
#gran vs tpa
prop.test(c(2090,2090),c(7406,13177))
```

###Diff comparisons


RA vs TPA

```{r}
ven_diff <- NULL
ven_diff$ra <- as.character(all$ens_gene_p[all$padj_r<0.05])
ven_diff$tpa <- as.character(all$ens_gene_p[all$padj_t<0.05])


V_diff <- Venn(ven_diff)
V_diff
pdf("../results/Venn_diff.pdf")
plot(V_diff, show=list(Faces=FALSE,DarkMatter=FALSE))
dev.off()
```

```{r}
ra <- unique(as.character(all$ens_gene_p[all$padj_r<0.05]))
tpa <- unique(as.character(all$ens_gene_p[all$padj_t<0.05]))
all_genes <- unique(as.character(all$ens_gene_p))
N <- 1000

inter.r <- length(intersect(ra,tpa))

getSimIntersect <- function(ra,tpa,all_genes){
  gA.sim <- sample(all_genes,length(ra))
  gB.sim <- sample(all_genes,length(tpa))
  return(length(intersect(gA.sim,gB.sim)))
}

inter.sim <- replicate(N,getSimIntersect(ra,tpa,all_genes))
p <- sum(inter.sim>=inter.r)/N
plot(density(inter.sim))
summary(inter.sim)
inter.r
quantile(inter.sim)
inter.r/mean(inter.sim)
```

RA vs TPA

```{r}
ra_tpa <- all_sig %>%
  filter(padj_ATRA<0.05 | padj_TPA<0.05) %>%
  mutate(Significant=case_when(padj_ATRA<0.05 & padj_TPA<0.05~"Both",
                               padj_ATRA<0.05 & padj_TPA>0.05~"ATRA-specific",
                               padj_ATRA>0.05 & padj_TPA<0.05~"TPA-specific"))

summary(lm(log2FoldChange_TPA~log2FoldChange_ATRA, data=ra_tpa))

ggplot(ra_tpa, aes(log2FoldChange_ATRA, log2FoldChange_TPA)) + geom_point(alpha=0.5) + ggformat + geom_hline(yintercept=0) + geom_vline(xintercept = 0) + geom_smooth(method="lm") + facet_wrap(~Significant, scales="free")
ggsave("../results/ATRA_vs_TPA_log2FC.pdf")

both_diff <- tidy(lm(log2FoldChange_TPA~log2FoldChange_ATRA, data=subset(all_sig, padj_ATRA<0.05 & padj_TPA<0.05)))
both_diff$model <- "ATRA and TPA"
ra <- tidy(lm(log2FoldChange_TPA~log2FoldChange_ATRA, data=subset(all_sig, padj_ATRA<0.05 & padj_TPA>0.05)))
ra$model <- "ATRA only"
tpa <- tidy(lm(log2FoldChange_TPA~log2FoldChange_ATRA, data=subset(all_sig, padj_ATRA>0.05 & padj_TPA<0.05)))
tpa$model <- "TPA only"

all_lm <- rbind(TNF_lm,both_diff,ra,tpa)
write.csv()
```
